==> 1. data_preparation.ipynb <==
# Data Preparation
This Jupyter notebook is designed to process data and map the binary positions of peptides within protein sequences.

## Getting the data

import pandas as pd
import numpy as np
import json
from Bio import SeqIO

# OsmitcStress.xlsx is the file that contains the data already filtered for the osmotic stress experiment
df = pd.read_excel('data/OsmoticStress.xlsx')

# Selecting specific columns from the dataset to simplify analysis
df = df[['Uniprot_ID', 'Peptide_sequence', 'Log2FC(LiP_norm)', 'Qvalue(LiP)']]

df.head()

# Reads data from the .Fasta file
def read_fasta_to_dict(file_path):
    """
    Reads a fasta file and returns a dictionary with the sequences.
    """
    sequences = {}
    for record in SeqIO.parse(file_path, "fasta"):
        uniprot_id = record.id.split('|')[1]
        sequences[uniprot_id] = str(record.seq)
    return sequences

fasta_file_path = 'data/UP000002311_559292.fasta'
sequences = read_fasta_to_dict(fasta_file_path)

# Map full protein sequences to each UniProt ID in the df
df["full_sequence"] = df['Uniprot_ID'].map(sequences)

# Fix the data types
df["full_sequence"]= df["full_sequence"].astype(str)
df["Peptide_sequence"]= df["Peptide_sequence"]. astype(str)

# Just get the sequences bellow length of 1000 to ease out memory usage
trimmed_df = df[df['full_sequence'].apply(lambda x: len(x)) < 1000]

def binary_position(row):
    """
    Generate a binary position list for a peptide sequence within its full sequence.
    """
    
    full_len = len(row['full_sequence'])
    short_len = len(row['Peptide_sequence'])
    start_index = row['full_sequence'].find(row['Peptide_sequence'])

    positions = np.zeros(full_len, dtype=int)

    if start_index != -1:
        positions[start_index:start_index + short_len] = 1
    return positions.tolist()

# Creation of the binary positions
trimmed_df['Binary_Positions'] = trimmed_df.apply(binary_position, axis=1)

max_length = 1000

# Padding binary position to length of 1000
# The padding is done by adding zeros to the right of the binary position list
# This is done to ensure that all binary position lists have the same length
trimmed_df['Padded_Binary_Positions'] = trimmed_df['Binary_Positions'].apply(lambda x: np.pad(x, (0, max_length - len(x)), 'constant'))

# Sort by bio relevance (Q value)
trimmed_df = trimmed_df.sort_values(by='Qvalue(LiP)', ascending=True)
trimmed_df.head()

# Get the 5000 most important peptides
trimmed_df = trimmed_df.head(5000)
trimmed_df.reset_index(inplace=True)

# Prevent formating problem
trimmed_df['Padded_Binary_Positions'] = trimmed_df['Padded_Binary_Positions'].apply(lambda x: json.dumps(x.tolist()))

# Save it to csv
trimmed_df.to_csv('data/OsmoticStress_with_binary_positions_padded_5000.csv', index=False)
