==> 2. create_embeddings.ipynb <==
# Embedding generation
This Jupyter notebook is designed to generate embeddings for each protein.

import esm
import torch
import pandas as pd
import tqdm
import tqdm
import numpy as np
import torch

# Loads the ESM-2 model
model, alphabet = esm.pretrained.esm2_t33_650M_UR50D()
model.eval()

# Get the sequence data already filtered
df = pd.read_csv('data/OsmoticStress_with_binary_positions_padded_5000.csv')

df_clean = df[["Uniprot_ID", "full_sequence", "Qvalue(LiP)"]]
df_clean.reset_index(drop=True, inplace=True)

# Fix type of full_sequence and remove NaN values
df_clean = df_clean[~df_clean['full_sequence'].apply(lambda x: isinstance(x, float))]
df_clean['full_sequence'] = df_clean['full_sequence'].astype(str)

# Retrieve sequences for unique protein IDs to avoid generating duplicate embeddings
df_clean_dropped = df_clean.drop_duplicates(subset='Uniprot_ID', keep='first')

df_clean_dropped

def generate_embeddings(model, alphabet, sequences):
    """
    Function to generate the embeddings.
    """
    embeddings = []
    
    batch_converter = alphabet.get_batch_converter()
    
    for sequence in tqdm.tqdm(sequences, desc="Generating Embeddings"):
        data = [(0, sequence)]
        batch_labels, batch_strs, batch_tokens = batch_converter(data)

        # Generate embeddings
        with torch.no_grad():
            results = model(batch_tokens, repr_layers=[33]) 
            token_embeddings = results['representations'][33]

        # Create full embeddings
        full_embedding = token_embeddings.squeeze().numpy()
        embeddings.append(full_embedding)
    
    return embeddings

def pad_sequence(sequence, target_length=1000, pad_token='<pad>'):
    """
    Append the <pad> token to the sequence until it reaches a length of 1000.
    """
    padding_length = target_length - len(sequence)
    return sequence + pad_token * padding_length

# Apply padding to each sequence with <pad>
df_clean_dropped['padded_sequence'] = df_clean_dropped['full_sequence'].apply(lambda x: pad_sequence(x))

# Check if the padding is correct
df_clean_dropped['full_sequence_length'] = df_clean_dropped['full_sequence'].apply(len)
df_clean_dropped['num_pad_tokens'] = df_clean_dropped['padded_sequence'].apply(lambda x: x.count('<pad>'))
df_clean_dropped['padded_sequence_length'] = df_clean_dropped['full_sequence_length'] + df_clean_dropped['num_pad_tokens']
df_clean_dropped['expected_padded_length'] = 1000

df_clean_dropped.reset_index(drop=True, inplace=True)
df_clean_dropped = df_clean_dropped[1000:]

# Generate the embeddings!
sequences = df_clean_dropped['padded_sequence'].tolist()
embeddings = generate_embeddings(model, alphabet, sequences)

df_test_dropped['full_embedding'] = [e.tolist() for e in embeddings]

# Save embeddings
df_test_dropped.to_pickle('data/embeddings_new_3_3.pkl', protocol=4)
