==> 2. create_embeddings_validation.ipynb <==
import esm
import torch
import pandas as pd
import tqdm

df = pd.read_excel('data/OsmoticStress.xlsx')

# Get a specific protein
df_P32485 = df[df["Uniprot_ID"] == "P32485"]

df_P32485

df_2 = pd.read_csv('data/OsmoticStress_with_binary_positions_padded.csv')
df_2 = df_2.drop(columns=['Binary_Positions', 'Padded_Binary_Positions'])
df_2_P32485 = df_2[df_2["Uniprot_ID"] == "P32485"]
df_2_P32485['full_sequence_len'] = df_2_P32485['full_sequence'].apply(len)

# Only get the P32485 HOG1
df_2_P32485_HOG1 = df_2_P32485.iloc[:1]
df_2_P32485_HOG1.head()

# Get the context of the peptide in the full sequence: 50 amino acids before and after the peptide
def get_peptide_context(full_seq, peptide, context_size=50):
    start_index = full_seq.find(peptide)
    if start_index == -1:
        return None, None, None  # Peptide not found
    start_slice = max(0, start_index - context_size)
    end_slice = min(len(full_seq), start_index + len(peptide) + context_size)
    return full_seq[start_slice:end_slice], start_slice, end_slice

# Apply the function to the first row of the dataframe
def apply_context(row):
    context, start_slice, end_slice = get_peptide_context(row['full_sequence'], row['Peptide_sequence'], context_size=50)
    return context, start_slice, end_slice

# Apply the function using the external lambda definition
df_context = df_2_P32485_HOG1.apply(apply_context, axis=1, result_type="expand")
df_context.columns = ['context_including_peptide', 'start_slice', 'end_slice']

df_2_P32485_HOG1

peptide_sequence = df_2_P32485_HOG1.iloc[0, 1]
print(len(peptide_sequence))
protein_sequence = df_2_P32485_HOG1.iloc[0, 3]
print(len(protein_sequence))

amino_acids = 'ACDEFGHIKLMNPQRSTVWY'

# Function to generate all mutant sequences
def generate_mutant_sequences(full_seq, context, start, end):
    mutants = []
    for index in range(len(context)):
        original_aa = context[index]
        for aa in amino_acids:
            if aa != original_aa:
                # Create the mutant sequence within the context
                mutant_context = context[:index] + aa + context[index + 1:]
                # Rebuild the full sequence with the mutated context
                mutant_sequence = full_seq[:start] + mutant_context + full_seq[end:]
                mutants.append((start + index, aa, mutant_sequence))
    return mutants

# Get all mutants
context, start, end = get_peptide_context(protein_sequence, peptide_sequence, 50)
mutants = generate_mutant_sequences(protein_sequence, context, start, end)

# Convert to DataFrame
df_mutants = pd.DataFrame(mutants, columns=['Position', 'Substituted_AA', 'Mutant_Sequence'])

df_mutants

def pad_sequence(sequence, target_length=1000, pad_token='<pad>'):
    """
    Add do the sequence <pad> to reach 1000.
    """
    padding_length = target_length - len(sequence)
    return sequence + pad_token * padding_length

# Apply padding to each sequence with <pad>
df_mutants['padded_sequence'] = df_mutants['Mutant_Sequence'].apply(lambda x: pad_sequence(x))

# Check if the padding is correct
df_mutants['full_sequence_length'] = df_mutants['Mutant_Sequence'].apply(len)
df_mutants['num_pad_tokens'] = df_mutants['padded_sequence'].apply(lambda x: x.count('<pad>'))
df_mutants['padded_sequence_length'] = df_mutants['full_sequence_length'] + df_mutants['num_pad_tokens']
df_mutants['expected_padded_length'] = 1000

df_mutants

# Loadsm the model
model, alphabet = esm.pretrained.esm2_t33_650M_UR50D()
model.eval()

import tqdm
import numpy as np
import torch

def generate_embeddings(model, alphabet, sequences):
    """
    Function to generate the embeddings.
    """
    embeddings = []
    
    batch_converter = alphabet.get_batch_converter()
    
    for sequence in tqdm.tqdm(sequences, desc="Generating Embeddings"):
        data = [(0, sequence)]
        batch_labels, batch_strs, batch_tokens = batch_converter(data)

        # Generate embeddings
        with torch.no_grad():
            results = model(batch_tokens, repr_layers=[33]) 
            token_embeddings = results['representations'][33]

        # Create full embeddings
        full_embedding = token_embeddings.squeeze().numpy()
        embeddings.append(full_embedding)
    
    return embeddings

df_clean = df_mutants[1000:]

sequences = df_clean['Mutant_Sequence'].tolist()
print(len(sequences))

# Generate the embeddings!
embeddings = generate_embeddings(model, alphabet, sequences)

df_clean['embeddings'] = [e.tolist() for e in embeddings]

# Save the data in a pickle file
df_clean.to_pickle('data/mutant_sequences_50_context_fix_2.pkl')
