==> 3. dl_model_data_prep.ipynb <==
# Deep Learning Model Data Preparation
This Jupyter notebook is dedicated to preparing the data for subsequent use in the deep learning model. It saves the data in pickle files to ensure efficient loading into the model in the next step, thereby avoiding memory load issues on the computer.

import json
import pickle
import pandas as pd
import numpy as np

def load_and_process_embedding(file_path):
    """
    Loads the embeddings pickle files.
    """
    with open(file_path, 'rb') as file:
        embeddings = pickle.load(file)
    embeddings = pd.DataFrame(embeddings)
    embeddings.set_index('Uniprot_ID', inplace=True)
    embeddings = embeddings.drop(columns=['full_sequence'])
    return embeddings

embeddings = pd.DataFrame()

embedding_files = [
    'data/embeddings_new_3_1.pkl',
    # 'data/embeddings_new_3_2.pkl',
    # 'data/embeddings_new_3_3.pkl',
]

for file_path in embedding_files:
    print(file_path)
    current_embeddings = load_and_process_embedding(file_path)
    embeddings = pd.concat([embeddings, current_embeddings])

embeddings_ = embeddings.reset_index()

# Creating trimmed embeddings: because of the first and last tokens
# (https://www.science.org/doi/suppl/10.1126/science.ade2574/suppl_file/science.ade2574_sm.pdf)
# We used BOS and EOS tokens to signal the beginning and end of a real protein, to allow the model to separate a full-sized protein from a cropped one.
trimmed_embeddings = [np.array(embedding)[1:-1, :] for embedding in embeddings['full_embedding']]
trimmed_embeddings_df = pd.DataFrame({'embeddings': trimmed_embeddings}, index=embeddings.index)

# We can see that we have the trimmed embedings ready to be used
trimmed_embeddings_df.head()

## Load Binary Position Data
Load the binary position data and corresponding target values from the CSV file created in notebook 1. This data is pre-processed, ensuring each binary position sequence is padded to a length of one thousand.

# Get the dataset
df = pd.read_csv('data/OsmoticStress_with_binary_positions_padded_5000.csv')
df['Binary_Positions'] = df['Binary_Positions'].apply(json.loads)
df['Padded_Binary_Positions'] = df['Padded_Binary_Positions'].apply(json.loads)
df.drop(columns=['full_sequence', 'Peptide_sequence', "Binary_Positions"], inplace=True)
df.set_index('Uniprot_ID', inplace=True)

df.head()

# Count the number of peptides per protein
df_ = df.reset_index()
counts = df_['Uniprot_ID'].value_counts()
print(counts)

### Prepare training dataset
Merge both embeddings and padded binary position together.

# Merge embeddings with the dataset with binary positions
dataset = df.merge(trimmed_embeddings_df, left_index=True, right_index=True, how='inner')
print(dataset.shape)

# Checks for missing values
dataset_ = dataset.reset_index()
counts = dataset_['Uniprot_ID'].value_counts()
print(counts)

dataset

# Saves the dataset to a pickle file
dataset.to_pickle('data/dataset_3.pkl')
